+++ apps/backend/src/api/tags.service.ts
@@ -1,10 +1,6 @@
 import { RepositoryService } from '../repository.service';
 
-import { HttpException, Injectable, Optional } from '@nestjs/common';
-import { SearchResult } from '../generated/openapi/model/searchResult';
-import { AbstractSearchLogicService } from 'src/searchlogic';
-import { SearchCandidate } from 'src/searchlogic';
-import { getSnapshotInfoFromId } from 'src/utils';
+import { HttpException, Injectable } from '@nestjs/common';
 
 @Injectable()
 export class TagsService {
