+++ apps/backend/src/controllers/files.controller.ts
@@ -1,5 +1,5 @@
-import { Controller, Get, Param, Query } from '@nestjs/common';
-import { FilesService } from '../api/files.service';
+import { Controller, Get, Param } from '@nestjs/common';
+import { FilesService } from '@/api/files.service';
 
 @Controller('files')
 export class FilesController {
