+++ apps/backend/src/controllers/registry.controller.ts
@@ -1,5 +1,5 @@
-import { Controller, Get, Param, Query } from '@nestjs/common';
-import { RegistryService } from '../api/registry.service';
+import { Controller, Get, Param } from '@nestjs/common';
+import { RegistryService } from '@/api/registry.service';
 
 @Controller('registry')
 export class RegistryController {
