+++ apps/backend/src/repository.service/dependency.impl.ts
@@ -2,7 +2,6 @@ import { PrismaClient } from '@prisma/client';
 import {
   fromThrowableSet,
   fromThrowableGet,
-  fromThrowableDelete,
 } from './fromThrowable';
 
 export async function setDependencies(
