+++ apps/backend/src/repository.service/index.ts
@@ -1,6 +1,5 @@
 import { Injectable, OnModuleDestroy, OnModuleInit } from '@nestjs/common';
 import { PrismaClient } from '@prisma/client';
-import { get } from 'http';
 
 import {
   getUser,
